+++ docs/about/whats_new.rst
@@ -15,6 +15,7 @@ v1.8rc (???)
 - Remove no longer required PyPEG2 dependency. (:pull:`840`)
 - Remove S3AIO driver. (:pull:`865`)
 - Change development version numbers generation. Use ``setuptools_scm`` instead of ``versioneer``. (:issue:`871`)
+- Deprecated ``datacube.helpers.write_geotiff``, use ``datacube.utils.cog.write_cog`` for similar functionality
 
 
 v1.7.0 (16 May 2019)